+++ python-wrapper/tests/test_gds.py
@@ -1,11 +1,24 @@
 import re
-from typing import Any
+from typing import Any, Generator
 
 import pandas as pd
 import pytest
 from pytest_mock import MockerFixture
 
 from neo4j_viz import Node
+
+
+@pytest.fixture(scope="class", autouse=True)
+def db_setup(gds: Any) -> Generator[None, None, None]:
+    gds.run_cypher(
+        "CREATE "
+        "  (a:_CI_A {name:'Alice', height:20, id:42, _id: 1337, caption: 'hello'})"
+        " ,(b:_CI_A:_CI_B {name:'Bob', height:10, id: 84, size: 11, labels: [1,2]})"
+        " ,(a)-[:KNOWS {year: 2025, id: 41, source: 1, target: 2}]->(b)"
+        " ,(b)-[:RELATED {year: 2015, _type: 'A', caption:'hej'}]->(a)"
+    )
+    yield
+    gds.run_cypher("MATCH (n:_CI_A|_CI_B) DETACH DELETE n")
 
 
 @pytest.mark.requires_neo4j_and_gds
@@ -72,6 +85,18 @@ def test_from_gds_integration_size(gds: Any) -> None:
             (1, 2, "REL2", "REL2", 2.0, 1.5),
             (2, 0, "REL", "REL", 3.0, 2.5),
         ]
+
+
+@pytest.mark.filterwarnings("ignore::DeprecationWarning")
+@pytest.mark.requires_neo4j_and_gds
+def test_from_gds_integration_all_db_properties(gds: Any, db_setup: None) -> None:
+    from neo4j_viz.gds import from_gds
+
+    with gds.graph.project("g2", ["_CI_A", "_CI_B"], "*") as G:
+        VG = from_gds(gds, G, node_radius_min_max=None, additional_db_node_properties=["name"])
+
+        assert len(VG.nodes) == 2
+        assert {n.properties["name"] for n in VG.nodes} == {"Alice", "Bob"}
 
 
 @pytest.mark.requires_neo4j_and_gds